requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
Pillow==10.2.0
firebase-admin==6.5.0
